requirements.txt
groq
protobuf
# Google Gemini API (stable client compatible with Gradio/Streamlit)
google-generativeai==0.7.2

# UI frameworks
streamlit==1.41.0
gradio==4.16.0

# General utilities
python-dotenv==1.0.1
pydantic==2.9.0

# Async/HTTP
httpx==0.27.2

# Data handling
pandas==2.2.3

# Optional scraping/text utils
beautifulsoup4==4.12.3
requests==2.32.3

# Config handling
pyyaml==6.0.2